Adds the @NewSpan and @SpanTag annotations
without this change there is no way to add local spans in an automated fashion. It's not also possible to add them to frameworks that provide only interfaces (such as Spring Data).
with this change we're adding this possibility.

This PR is an adoption of @Koizumi85 work started here - https://github.com/Koizumi85/spring-cloud-sleuth-annotation

fixes #182