TraceFilter skipPattern and extra span when X-B3-Flags header is present
without this change if there's a debug flag we set traceid & spanid for the first request inside the zipkin http extractor. we don't set the parent span flag so an additional span is created.
with this change we set the trace and span id ONLY when the debug flag is set AND the request is malformed. That means - only when the span id is set and there is no trace id.

fixes #665