Continuation of async spans (#784)
without this change, we always create a new span in TraceAsyncAspect. That's because the span created via LazyTraceExecutor (which is executed before the aspect), contains no useful information.
with this change we're renaming the LTE span in TraceAsyncAspect. We're using an InternalApi class that we kindly ask that nobody uses.

fixes #661 #783