onError & onComplete should reuse span from ThreadLocal
we've played around with the working M7 version, and only after rolling back the changes on wrapping of onError and onComplete with try-with-resources, did we manage to successfully propagate the span context to various places. We're not really sure if that's healthy so we're pushing this to a branch for further tests.

fixes gh-973