fix #994 Make SpanSubscriber fake fuseability
This is done by implementing NO-OP QueueSubscription on the subscriber
and removing the guard against `Fuseable` publisher sources.

As `Fuseable` is a marker interface there should be no reason for code
to perform casts so for now the fact that the `lift` call in
ReactorSleuth doesn't actually produces a `Fuseable` publisher should
not be a problem.