Wrapping the Feign client only when using the LoadBalancerFeignClient
without this change we wrapped in the handleSend and handleReceive both the method wrapping the execution of the TraceLoadBalancerFeignClient and its delegate too. That way we had 2 pairs of CS & CR.
with this change we disable that functionality by wrapping only the delegate. The problem with such an approach was such that there were exceptions thrown before the delegate was executed. That's why we added a fallback mechanism in that case.

fixes gh-1007