Brought back the onEachOperator logic
we wanted to improve the performance of wrapping reactor by introducing refactoring that used onLastOperator instead of onEach. ALthough the performance was improved, the code stopped working the way it should.
with this commit, we're coming back to making the feature work again but the performance might decline slightly.

fixes gh-1204
fixes gh-1147
fixes gh-1143