Makes a nice toString on RestTemplateSender
The AsyncZipkinSpanHandler calls 'check' once on startup to let someone
know an error that may affect tracing up front. Before, this didn't
include the endpoint so it is less obvious what could be the problem.

Ex people goof the URL (don't add /api/v2/spans or it is added twice)