Fixing #1797 via wiring beans by id not by type.
It seems M6 introduced two dependencies in TraceWebSocketAutoConfiguration:
- Propagator.Setter<MessageHeaderAccessor>
- Propagator.Getter<MessageHeaderAccessor>

Both of them are created in TraceSpringMessagingAutoConfiguration. If you look closer, they are referencing the very same object: MessageHeaderPropagation.INSTANCE
This is possible because MessageHeaderPropagation implements both Propagator.Setter<MessageHeaderAccessor> and Propagator.Getter<MessageHeaderAccessor>. So when the context is created there will be two beans:
- traceMessagePropagationSetter
- traceMessagePropagationGetter

Both of them are a Propagator.Setter<MessageHeaderAccessor> and a Propagator.Getter<MessageHeaderAccessor> at the same time. So we have two beans both of them are getters and setters so we have two getters and two setters at runtime and Spring does not know which one is the right one, one possible fix is wiring these beans in using their id.