Fixed W3C tracestate and baggage setting
without this change we're not setting tracestate whatsover. Baggage is also not being propagated properly (via the baggage field)
with this change when tracestate is there in the headers it will be treated as baggage. When the context is propagated then tracestate will not be put to the baggage field

fixes gh-2001